+++ packages/components/src/Datalist/Datalist.component.test.js
@@ -1,6 +1,8 @@
 import React from 'react';
-import { render, screen, fireEvent } from '@testing-library/react';
+
+import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
+
 import Datalist from './Datalist.component';
 
 const props = {
@@ -272,6 +274,24 @@ describe('Datalist component', () => {
 
 		// then
 		expect(screen.getByRole('textbox')).toHaveValue('My foo updated');
+	});
+
+	it('should keep filter when titleMap is updated', async () => {
+		// given
+		const testProps = {
+			id: 'my-datalist',
+			value: 'foo',
+			onChange: jest.fn(),
+			...props,
+		};
+		const { rerender } = render(<Datalist {...testProps} />);
+
+		// when
+		userEvent.type(screen.getByRole('textbox'), 'a');
+		rerender(<Datalist {...testProps} titleMap={[...props.titleMap]} />);
+
+		// then
+		expect(screen.getByRole('textbox')).toHaveValue('a');
 	});
 
 	it('should set highlight on current value suggestion', () => {