+++ packages/containers/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @talend/react-containers
 
+## 7.2.2
+
+### Patch Changes
+
+- Updated dependencies [9a581a4bc]
+- Updated dependencies [c8fbf0130]
+- Updated dependencies [1160ec1f6]
+  - @talend/react-components@7.0.0
+  - @talend/react-forms@7.3.1
+
 ## 7.2.1
 
 ### Patch Changes